+++ src/internals/observables.ts
@@ -7,8 +7,3 @@ export type ScreenSize = 'small' | 'medium' | 'large';
  * 현재 화면 크기 상태 
  */
 export const screen: IObservableValue<ScreenSize> = observable.box('large');
-
-/**
- * 로딩 진행 상태 (0 - 100) 
- */
-export const progress: IObservableValue<number> = observable.box(0);